+++ src/main/java/com/WalkiePaw/presentation/domain/report/boardReportDto/request/BoardReportAddRequest.java
@@ -12,8 +12,8 @@ import lombok.ToString;
 public class BoardReportAddRequest {
     private BoardReportCategory reason;
     private String content;
-    private Integer memberId;
-    private Integer boardId;
+    private Long memberId;
+    private Long boardId;
 
     /**
      * DTO -> Entity
